Cleaned up file reading hierarchy
