+++ src/CMakeLists.txt
@@ -23,13 +23,13 @@ add_subdirectory(input)
 if (USE_PYTHON)
   # AEStream Python
   add_subdirectory(pybind)
-else() 
-  # AEStream executable
-  add_executable(aestream aestream.cpp)
-  target_include_directories(aestream PRIVATE "${TORCH_INCLUDE_DIRS}" ${CMAKE_INSTALL_LIBDIR})
-  target_link_libraries(aestream PRIVATE aer aestream_file aestream_input aestream_output)
+endif() 
 
-  # Install targets
-  # install(TARGETS aer aestream_input aestream_output aestream_lib LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR})
-  install(IMPORTED_RUNTIME_ARTIFACTS aestream DESTINATION ${CMAKE_INSTALL_BINDIR})
-endif()
+# AEStream executable
+add_executable(aestream aestream.cpp)
+target_include_directories(aestream PRIVATE ${CMAKE_INSTALL_LIBDIR})
+target_link_libraries(aestream PRIVATE aer aestream_file aestream_input aestream_output)
+
+# Install targets
+# install(TARGETS aer aestream_input aestream_output aestream_lib LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR})
+install(IMPORTED_RUNTIME_ARTIFACTS aestream DESTINATION ${CMAKE_INSTALL_BINDIR})
