+++ src/file/utils.hpp
@@ -4,27 +4,24 @@
 #include <queue>
 #include <string>
 
+#include "../aer.hpp"
+#include "../generator.hpp"
+
+struct FileBase {
+  virtual ~FileBase() = default;
+  virtual Generator<AER::Event> stream(size_t n_events = -1) = 0;
+  virtual std::tuple<AER::Event *, size_t>
+  read_events(const size_t &n_events = -1) = 0;
+};
+
 void close_file(FILE *fp);
 
 // typedef std::unique_ptr<FILE, decltype(&close_file)> unique_file_t;
-typedef std::shared_ptr<FILE> shared_file_t;
+typedef std::unique_ptr<FILE, decltype(&close_file)> file_t;
 
-shared_file_t open_file(const std::string &filename);
-
-template <typename T, typename H, H h(const shared_file_t &),
-          T f(const shared_file_t &, const H &, uint64_t *)>
-T file_reader(const std::string &filename, const H &header);
+file_t open_file(const std::string &filename);
 
 bool ends_with(std::string const &value, std::string const &ending);
-
-// template <typename T, typename R>
-// std::queue<T> file_to_parts(const unique_file_t &fp, const size_t &size,
-//                             const R &header);
-
-// template <typename T, typename R>
-// std::queue<T> file_to_intervals(const unique_file_t &fp, const size_t
-// &interval,
-//                                 const R &header);
 
 #define HEADER_START 0x25
 #define HEADER_END 0x0A
