+++ src/input/file.cpp
@@ -1,14 +1,18 @@
 #include "file.hpp"
 
-Generator<AER::Event> file_event_generator(const std::string filename,
-                                           const std::atomic<bool> &runFlag) {
-  auto fp = open_file(filename);
+std::unique_ptr<FileBase> file_base(const std::string &filename) {
+  const auto fp = open_file(filename);
 
   if (ends_with(filename, ".dat")) {
-    return dat_stream_events(filename);
+    return std::unique_ptr<FileBase>(new DAT(fp));
   } else if (ends_with(filename, ".aedat4")) {
-    return AEDAT4(fp).stream();
+    return std::unique_ptr<FileBase>(new AEDAT4(fp));
   } else {
     throw std::invalid_argument("Unknown file type " + filename);
   }
+}
+
+Generator<AER::Event> file_event_generator(const std::string filename,
+                                           const std::atomic<bool> &runFlag) {
+  return file_base(filename)->stream(-1);
 }
